requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
drf-access-policy==1.3.0
django-otp==1.1.4
django-storages==1.13.2
django-filter==21.1
django-notifications-hq==1.7.0
django-cleanup==7.0.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
celery[SQLAlchemy]==5.2.7
PyJWT==2.6.0
psycopg2==2.9.5
gunicorn==20.1.0
uvicorn==0.20.0
drf-yasg==1.21.5
Pillow==9.4.0
boto3==1.26.70
qrcode==7.4.2
Cerberus==1.3.4
asgiref==3.6.0
python-dateutil==2.8.2